Renames components, eliminates unused controllers folder
