--- src/components/controllers/navbar.js
@@ -1,6 +1,0 @@
-const navbar = () => {
-        
-    return null
-}
-
-export default navbar;
